+++ hardware/X6-1000M/src/X6ADC/libx6adc.h
@@ -53,7 +53,7 @@ EXPORT int set_averager_settings(int, int, int, int, int);
 EXPORT int acquire(int);
 EXPORT int wait_for_acquisition(int, int);
 EXPORT int stop(int);
-EXPORT int transfer_waveform(int, int, short *, size_t);
+EXPORT int transfer_waveform(int, int, short *, unsigned int);
 
 EXPORT int set_log(char *);
 int update_log(FILE * pFile);
